+++ crates/miden-lib/src/account/interface/test.rs
@@ -12,7 +12,7 @@ use miden_objects::account::{
 };
 use miden_objects::assembly::diagnostics::NamedSource;
 use miden_objects::assembly::{Assembler, DefaultSourceManager};
-use miden_objects::asset::{FungibleAsset, NonFungibleAsset, TokenSymbol};
+use miden_objects::asset::{FungibleAsset, NonFungibleAsset, TokenName, TokenSymbol};
 use miden_objects::crypto::rand::{FeltRng, RpoRandomCoin};
 use miden_objects::note::{
     Note,
@@ -68,6 +68,8 @@ fn test_basic_wallet_default_notes() {
                 TokenSymbol::new("POL").expect("invalid token symbol"),
                 10,
                 Felt::new(100),
+                TokenName::new("polygon").expect("invalid token name"),
+                None,
             )
             .expect("failed to create a fungible faucet component"),
         )
@@ -340,6 +342,8 @@ fn test_basic_fungible_faucet_custom_notes() {
                 TokenSymbol::new("POL").expect("invalid token symbol"),
                 10,
                 Felt::new(100),
+                TokenName::new("polygon").expect("Invalid token name"),
+                None,
             )
             .expect("failed to create a fungible faucet component"),
         )